requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.11.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
whitenoise==5.0.1
